Part of #2086: * copied NewSubmissionWizard.js into SubmissionWizard.js * started replacing tests * installed @testing-library/react as enzyme doesn't work with hooks yet
