+++ package.json
@@ -152,6 +152,7 @@
     "yup": "^0.26.2"
   },
   "devDependencies": {
+    "@testing-library/react": "^8.0.1",
     "@yarnpkg/lockfile": "^1.1.0",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.0.2",
