+++ packages/component-submission/client/index.js
@@ -1,3 +1,2 @@
 export { default as SubmissionWizard } from './pages/SubmissionWizard'
-export { default as NewSubmissionWizard } from './pages/NewSubmissionWizard'
 export { default as ThankYouPage } from './pages/ThankYouPage'
