+++ packages/component-submission/client/pages/SubmissionWizard.js
@@ -1,25 +1,36 @@
-import React from 'react'
+import React, { useState } from 'react'
+import { compose, branch, renderComponent } from 'recompose'
 import { Switch, Redirect } from 'react-router-dom'
-import {
-  compose,
-  withHandlers,
-  withProps,
-  branch,
-  renderComponent,
-} from 'recompose'
+import { Box, Flex } from '@rebass/grid'
+import styled from 'styled-components'
+import { th } from '@pubsweet/ui-toolkit'
+import { Formik, Form } from 'formik'
+import * as yup from 'yup'
 import {
   ErrorPage,
   TrackedRoute,
 } from '@elifesciences/component-elife-app/client'
 import { Loading } from '@elifesciences/component-elife-ui/client/atoms'
+import { Button } from '@pubsweet/ui'
 
-import WizardStep from '../components/WizardStep'
+import AuthorStep from './AuthorStepPage'
+import FilesStep from './FilesStepPage'
+import DetailsStep from './DetailsStepPage'
+import EditorStep from './EditorsStepPage'
+import DisclosureStep from './DisclosureStepPage'
 
-import AuthorStepPage from './AuthorStepPage'
-import DisclosureStepPage from './DisclosureStepPage'
-import FilesStepPage from './FilesStepPage'
-import EditorsStepPage from './EditorsStepPage'
-import DetailsStepPage from './DetailsStepPage'
+import SubmissionSave from '../components/SubmissionSave'
+import WizardSubmit from '../components/WizardSubmit'
+import ProgressBar from '../components/ProgressBar'
+import wizardWithGQL from '../graphql/wizardWithGQL'
+import {
+  parseInputToFormData,
+  parseFormToOutputData,
+  flattenObject,
+  getErrorStepsFromErrors,
+  convertArrayToReadableList,
+} from '../utils'
+import { STEP_NAMES } from '../utils/constants'
 
 import wizardSchema, {
   authorSchema,
@@ -28,150 +39,200 @@ import wizardSchema, {
   editorsSchema,
 } from '../utils/ValidationSchemas'
 
-import wizardWithGQL from '../graphql/wizardWithGQL'
+const BoxNoMinWidth = styled(Box)`
+  min-width: 0;
+`
 
-import { parseFormToOutputData, parseInputToFormData } from '../utils'
+const ErrorMessage = styled.div`
+  color: ${th('colorError')};
+`
 
-export class SubmissionWizard extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      suspendSave: false,
-      isUploading: props.initialValues.fileStatus === 'CHANGING',
-      nextSaveValues: null,
-    }
-  }
-
-  disableSave = saveStatus => this.setState({ suspendSave: saveStatus })
-
-  setIsUploading = status => {
-    // status: boolean
-    this.disableSave(status)
-    this.setState({ isUploading: status })
-    if (!status && this.state.nextSaveValues) {
-      this.onNextClick(this.state.nextSaveValues)
-      this.setState({ nextSaveValues: null })
-    }
-  }
-
-  onNextClick = values => {
-    if (!this.state.suspendSave && !this.state.isUploading) {
-      return this.props.updateManuscript(values)
-    }
-    this.setState({ nextSaveValues: values })
-    return Promise.resolve() // this seems hacky and should probably be changed
-  }
-
-  onSubmitManuscript = async values => {
-    this.disableSave(true)
-    this.props.submitManuscript(values).finally(() => {
-      this.disableSave(false)
-    })
-  }
-
-  render() {
-    const { match, history, initialValues } = this.props
-
-    return (
-      <Switch>
-        <TrackedRoute
-          path={`${match.path}/author`}
-          render={() => (
-            <WizardStep
-              component={AuthorStepPage}
-              handleAutoSave={this.onNextClick}
-              handleButtonClick={this.onNextClick}
-              history={history}
-              initialValues={initialValues}
-              nextUrl={`${match.url}/files`}
-              step={0}
-              title="Your details"
-              validationSchema={authorSchema}
-            />
-          )}
-        />
-        <TrackedRoute
-          exact
-          path={`${match.path}/files`}
-          render={() => (
-            <WizardStep
-              component={FilesStepPage}
-              disableSave={this.disableSave}
-              handleAutoSave={this.onNextClick}
-              handleButtonClick={this.onNextClick}
-              history={history}
-              initialValues={initialValues}
-              isUploading={this.state.isUploading}
-              nextUrl={`${match.url}/details`}
-              previousUrl={`${match.url}/author`}
-              setIsUploading={this.setIsUploading}
-              step={1}
-              validationSchema={filesSchema}
-            />
-          )}
-        />
-        <TrackedRoute
-          path={`${match.path}/details`}
-          render={() => (
-            <WizardStep
-              component={DetailsStepPage}
-              handleAutoSave={this.onNextClick}
-              handleButtonClick={this.onNextClick}
-              history={history}
-              initialValues={initialValues}
-              nextUrl={`${match.url}/editors`}
-              previousUrl={`${match.url}/files`}
-              step={2}
-              title="Help us get your work seen by the right people"
-              validationSchema={submissionSchema}
-            />
-          )}
-        />
-        <TrackedRoute
-          path={`${match.path}/editors`}
-          render={() => (
-            <WizardStep
-              component={EditorsStepPage}
-              handleAutoSave={this.onNextClick}
-              handleButtonClick={this.onNextClick}
-              history={history}
-              initialValues={initialValues}
-              nextUrl={`${match.url}/disclosure`}
-              previousUrl={`${match.url}/details`}
-              step={3}
-              title="Who should review your work?"
-              validationSchema={editorsSchema}
-            />
-          )}
-        />
-        <TrackedRoute
-          path={`${match.path}/disclosure`}
-          render={() => (
-            <WizardStep
-              component={DisclosureStepPage}
-              finalStep
-              handleAutoSave={this.onNextClick}
-              handleButtonClick={this.onSubmitManuscript}
-              history={history}
-              initialValues={initialValues}
-              nextUrl={`/thankyou/${match.params.id}`}
-              previousUrl={`${match.url}/editors`}
-              step={4}
-              title="Disclosure of data to editors"
-              validationSchema={wizardSchema}
-            />
-          )}
-        />
-        <Redirect from="/submit/:id" to={`/submit/${match.params.id}/author`} />
-        <ErrorPage error="404: page not found" />
-      </Switch>
+export const SubmissionWizard = ({
+  data,
+  match,
+  history,
+  updateManuscript,
+  submitManuscript,
+}) => {
+  const getCurrentStepFromPath = () =>
+    STEP_NAMES.map(step => step.toLowerCase()).indexOf(
+      history.location.pathname.split('/')[3].toLowerCase(),
     )
-  }
+
+  const [currentStep, setCurrentStep] = useState(getCurrentStepFromPath())
+  const [isUploading, setIsUploading] = useState(false)
+  const [submissionAttempted, setsubmissionAttempted] = useState(false)
+
+  const isLastStep = () => currentStep === STEP_NAMES.length - 1
+  const initialValues = parseInputToFormData(data.manuscript)
+  const stepValidation = [
+    authorSchema,
+    filesSchema,
+    submissionSchema,
+    editorsSchema,
+    wizardSchema,
+  ]
+
+  const handleSave = formValues =>
+    updateManuscript({
+      variables: { data: parseFormToOutputData(formValues) },
+    })
+
+  const handleSubmit = formValues =>
+    submitManuscript({
+      variables: { data: parseFormToOutputData(formValues) },
+    })
+
+  return (
+    <Formik
+      initialValues={initialValues}
+      onSubmit={handleSubmit}
+      render={formikProps => (
+        <Form data-testid="submission-wizard-form">
+          <SubmissionSave
+            disabled={formikProps.isSubmitting}
+            handleSave={handleSave}
+            values={{
+              ...formikProps.values,
+              lastStepVisited: history.location.pathname,
+            }}
+          />
+          <Flex>
+            <BoxNoMinWidth flex="1 1 100%" mx={[0, 0, 0, '16.666%']}>
+              <Box my={5}>
+                <ProgressBar currentStep={currentStep} />
+              </Box>
+              <Switch>
+                <TrackedRoute
+                  path={`${match.path}/author`}
+                  render={() => <AuthorStep {...formikProps} />}
+                />
+                <TrackedRoute
+                  path={`${match.path}/files`}
+                  render={() => (
+                    <FilesStep
+                      {...formikProps}
+                      isUploading={isUploading}
+                      setIsUploading={setIsUploading}
+                    />
+                  )}
+                />
+                <TrackedRoute
+                  path={`${match.path}/details`}
+                  render={() => <DetailsStep {...formikProps} />}
+                />
+                <TrackedRoute
+                  path={`${match.path}/editors`}
+                  render={() => <EditorStep {...formikProps} />}
+                />
+                <TrackedRoute
+                  path={`${match.path}/disclosure`}
+                  render={() => (
+                    <DisclosureStep
+                      isSubmissionAttempted={submissionAttempted}
+                      {...formikProps}
+                    />
+                  )}
+                />
+                <Redirect
+                  from="/newSubmit/:id"
+                  to={`/newSubmit/${match.params.id}/author`}
+                />
+                <ErrorPage error="404: page not found" />
+              </Switch>
+              {!!Object.keys(formikProps.errors).length &&
+                submissionAttempted &&
+                isLastStep() && (
+                  <ErrorMessage data-test-id="test-error-message">
+                    We&apos;re sorry but there appears to be one or more errors
+                    in your submission that require attention before you can
+                    submit. Please use the back button to review the{' '}
+                    {convertArrayToReadableList(
+                      getErrorStepsFromErrors(formikProps.errors),
+                    )}{' '}
+                    steps before trying again.
+                  </ErrorMessage>
+                )}
+              <Flex mt={6}>
+                <Box mr={3}>
+                  <Button
+                    data-test-id="back"
+                    disabled={currentStep === 0}
+                    onClick={() => {
+                      setCurrentStep(currentStep - 1)
+                      history.push(
+                        `${match.url}/${STEP_NAMES[
+                          currentStep - 1
+                        ].toLowerCase()}`,
+                      )
+                    }}
+                  >
+                    Back
+                  </Button>
+                </Box>
+                <Box>
+                  {currentStep === STEP_NAMES.length - 1 ? (
+                    <WizardSubmit
+                      setSubmissionAttempted={setsubmissionAttempted}
+                      setTouched={formikProps.setTouched}
+                      submitForm={formikProps.submitForm}
+                      validateForm={formikProps.validateForm}
+                    />
+                  ) : (
+                    <Button
+                      data-test-id="next"
+                      onClick={() => {
+                        formikProps.validateForm().then(errors => {
+                          if (!Object.keys(errors).length) {
+                            setCurrentStep(currentStep + 1)
+                            history.push(
+                              `${match.url}/${STEP_NAMES[
+                                currentStep + 1
+                              ].toLowerCase()}`,
+                            )
+                          }
+
+                          Object.keys(errors).forEach(errorField => {
+                            if (typeof errors[errorField] === 'object') {
+                              const flattenedSubfields = flattenObject(
+                                errors[errorField],
+                              )
+                              Object.keys(flattenedSubfields).forEach(
+                                subField => {
+                                  formikProps.setFieldTouched(
+                                    `${errorField}.${subField}`,
+                                    true,
+                                  )
+                                },
+                              )
+                            } else {
+                              formikProps.setFieldTouched(
+                                errorField,
+                                true,
+                                false,
+                              )
+                            }
+                          })
+                        })
+                      }}
+                      primary
+                    >
+                      Next
+                    </Button>
+                  )}
+                </Box>
+              </Flex>
+            </BoxNoMinWidth>
+          </Flex>
+        </Form>
+      )}
+      validationSchema={yup.object().shape(stepValidation[currentStep])}
+    />
+  )
 }
 
 export default compose(
   wizardWithGQL,
-  // TODO: This error / loading branch wrapper is a common thing across app. Should be turned into an import
   branch(
     props => props.data && (props.data.loading && !props.data.manuscripts),
     renderComponent(Loading),
@@ -181,17 +242,4 @@ export default compose(
     props => props.data.manuscript.clientStatus !== 'CONTINUE_SUBMISSION',
     () => () => <Redirect to="/" />,
   ),
-  withProps(props => ({
-    initialValues: parseInputToFormData(props.data.manuscript),
-  })),
-  withHandlers({
-    updateManuscript: props => formValues =>
-      props.updateManuscript({
-        variables: { data: parseFormToOutputData(formValues) },
-      }),
-    submitManuscript: props => formValues =>
-      props.submitManuscript({
-        variables: { data: parseFormToOutputData(formValues) },
-      }),
-  }),
 )(SubmissionWizard)
